Generalize recent transaction components
